correctly paginate through events
The issue with pagination based on limit and offset was that events
might change their status (from Failed to Processed) during event
consumption. This resulted in an incorrect offset. Instead of fixed
pages based on `offset`, we get the next `limit` events with `arena_id`
greater than the last processed event.